backend/requirements.txt
Django
psycopg2
graphene_django
flake8
factory_boy
django-cors-headers